drawer page fixed and splitted to different drawer pages, tests splitted based on new pages
