+++ jdi-light-material-ui-tests/src/main/java/io/github/com/StaticSite.java
@@ -34,15 +34,7 @@ import io.github.com.pages.layout.BoxPage;
 import io.github.com.pages.layout.ContainerPage;
 import io.github.com.pages.layout.GridPage;
 import io.github.com.pages.layout.HiddenPage;
-import io.github.com.pages.navigation.BottomNavigationPage;
-import io.github.com.pages.navigation.ContextMenuPage;
-import io.github.com.pages.navigation.DrawerPage;
-import io.github.com.pages.navigation.LinkPage;
-import io.github.com.pages.navigation.RouterBreadcrumbsPage;
-import io.github.com.pages.navigation.SimpleBreadcrumbsPage;
-import io.github.com.pages.navigation.SimpleMenuPage;
-import io.github.com.pages.navigation.StepperPage;
-import io.github.com.pages.navigation.TabPage;
+import io.github.com.pages.navigation.*;
 import io.github.com.pages.surfaces.AccordionPage;
 import io.github.com.pages.surfaces.AppBarPage;
 import io.github.com.pages.surfaces.BackToTopPage;
@@ -281,25 +273,25 @@ public class StaticSite {
 
     // drawer
     @Url("/temporary_drawer")
-    public static DrawerPage temporaryDrawerPage;
+    public static TemporaryDrawerPage temporaryDrawerPage;
 
     @Url("/swipeable_drawer")
-    public static DrawerPage swipeableDrawerPage;
+    public static SwipeableDrawerPage swipeableDrawerPage;
 
     @Url("/responsive_drawer")
-    public static DrawerPage responsiveDrawerPage;
+    public static ResponsiveDrawerPage responsiveDrawerPage;
 
     @Url("/persistent_drawer")
-    public static DrawerPage persistentDrawerPage;
+    public static PersistentDrawerPage persistentDrawerPage;
 
     @Url("/mini_drawer")
-    public static DrawerPage miniDrawerPage;
+    public static MiniDrawerPage miniDrawerPage;
 
     @Url("/permanent_drawer")
-    public static DrawerPage permanentDrawerPage;
+    public static PermanentDrawerPage permanentDrawerPage;
 
     @Url("/clipped_drawer")
-    public static DrawerPage clippedDrawerPage;
+    public static ClippedDrawerPage clippedDrawerPage;
 
     // stepper
